evening19.03a

+++ Patients/src/main/java/com/chvei/DoP/services/servicesImp/VisitServiceImp.java
@@ -4,10 +4,12 @@ import com.chvei.DoP.DTO.VisitDTO;
 import com.chvei.DoP.entity.Patient;
 import com.chvei.DoP.entity.Visit;
 import com.chvei.DoP.exceptions.ResourceNotFoundException;
+import com.chvei.DoP.exceptions.UnacceptableActionException;
 import com.chvei.DoP.repositories.VisitRepository;
 import com.chvei.DoP.services.PatientService;
 import com.chvei.DoP.services.VisitService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -17,6 +19,7 @@ import java.util.logging.Logger;
 @Service
 public class VisitServiceImp implements VisitService {
     private final Logger logger = Logger.getLogger(VisitServiceImp.class.getName());
+    private final Sort descending = Sort.by(Sort.Direction.DESC, "created");
     private VisitRepository visitRepository;
     private PatientService patientService;
 
@@ -30,8 +33,21 @@ public class VisitServiceImp implements VisitService {
     }
 
     @Override
+    public int getCountByPatientID(Long id) {
+        return visitRepository.countAllByPatient_Id(id);
+    }
+
+    @Override
+    public List<VisitDTO> getFirstTenVisitsByPatientIdDesc(Long id) {
+        return visitRepository.findFirst10ByPatient_IdOrderByCreatedDesc(id)
+                .stream()
+                .map(this::toDTO)
+                .toList();
+    }
+
+    @Override
     public List<VisitDTO> getVisitsByPatientId(Long id) {
-        return visitRepository.findByPatient_Id(id)
+        return visitRepository.findByPatient_Id(id, descending)
                 .stream()
                 .map(this::toDTO)
                 .toList();
@@ -54,6 +70,9 @@ public class VisitServiceImp implements VisitService {
 
     @Override
     public VisitDTO saveVisit(VisitDTO visitDTO) {
+        if (visitRepository.existsByCreatedAndPatient_Id(visitDTO.getCreated(), visitDTO.getPatientId())) {
+            throw new UnacceptableActionException("На дату " + visitDTO.getCreated().toString() + " визит уже существует");
+        }
         Patient patient = patientService.getPatientById(visitDTO.getPatientId());
         Visit visit = visitRepository.save(toEntity(visitDTO));
         patient.addVisit(visit);
@@ -74,13 +93,20 @@ public class VisitServiceImp implements VisitService {
     }
 
     @Override
-    public void deleteVisit(Long id) {
+    public boolean deleteVisit(Long id) {
         Visit visit = visitRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Visit with id " + id + " not found"));
+        if (visit.isRegistration()) {
+            throw new UnacceptableActionException("The visit cannot be deleted because he is responsible for registering the patient");
+        }
         Patient patient = patientService.getPatientById(visit.getPatient().getId());
         patient.removeVisit(visit);
         visitRepository.deleteById(id);
-        logger.log(Level.INFO, "Visit to " + patient.getLastName() + " " + visit.getCreated() + " delete");
+        boolean delete = visitRepository.existsById(id);
+        if (!delete) {
+            logger.log(Level.INFO, "Visit to " + patient.getLastName() + " " + visit.getCreated() + " delete");
+        }
+        return delete;
     }
 
     public VisitDTO toDTO(Visit visit) {
@@ -89,6 +115,7 @@ public class VisitServiceImp implements VisitService {
         visitDTO.setText(visit.getText());
         visitDTO.setPatientId(visit.getPatient().getId());
         visitDTO.setCreated(visit.getCreated());
+        visitDTO.setRegistration(visit.isRegistration());
         return visitDTO;
     }
 
@@ -98,6 +125,7 @@ public class VisitServiceImp implements VisitService {
         visit.setCreated(visitDTO.getCreated());
         visit.setText(visitDTO.getText());
         visit.setPatient(patientService.getPatientById(visitDTO.getPatientId()));
+        visit.setRegistration(visit.isRegistration());
         return visit;
     }
 }
